+++ CHANGELOG.md
@@ -6,6 +6,13 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
 
+## [0.3.0] - 2024-08-12
+- support float indices e.g. "(Mg0.4Fe1.6)2.0SiO4" == Olivine
+- add minerals.h (very short list)
+- update atomic_weight_formula.ino
+
+----
+
 ## [0.2.2] - 2024-04-07
 - update GitHub actions
 - refactor internal names for readability